feat(backend): Implement user-specific board creation and enhance column management with user authentication

+++ apps/backend/src/column/column.service.ts
@@ -1,4 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ForbiddenException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { ColumnInput } from './dto/column.dto';
 import { UUID } from 'crypto';
 import { PrismaService } from '../prisma/prisma.service';
@@ -7,18 +11,9 @@ import { PrismaService } from '../prisma/prisma.service';
 export class ColumnService {
   constructor(private readonly prisma: PrismaService) {}
 
-  async create(column: ColumnInput) {
-    return this.prisma.column.create({
-      data: {
-        title: column.title,
-        order: column.order || 0,
-      },
-    });
-  }
-
-  async findAll() {
+  async findAll(userId: number) {
     return this.prisma.column.findMany({
-      where: {},
+      where: { board: { userId } },
       include: {
         tasks: {
           orderBy: { order: 'asc' },
@@ -27,9 +22,39 @@ export class ColumnService {
     });
   }
 
+  async create(column: ColumnInput, userId: number) {
+    const userBoard = await this.prisma.board.findFirst({
+      where: { userId: userId },
+    });
+
+    if (!userBoard) {
+      throw new NotFoundException('Kullanıcıya ait bir pano bulunamadı.');
+    }
+
+    return this.prisma.column.create({
+      data: {
+        title: column.title,
+        boardId: userBoard.id,
+      },
+    });
+  }
+
   async update() {}
 
-  async remove(uuid: UUID) {
+  async remove(uuid: UUID, userId: number) {
+    const column = await this.prisma.column.findUnique({
+      where: { uuid: uuid },
+      include: { board: true },
+    });
+
+    if (!column) {
+      throw new NotFoundException('Kolon bulunamadı.');
+    }
+
+    if (column.board.userId !== userId) {
+      throw new ForbiddenException('Bu kolonu silme yetkiniz yok.');
+    }
+
     return this.prisma.column.delete({
       where: { uuid: uuid },
     });
